Fixed bugs: Sending empty age or height to calorie calories, missing daily goal value from DB on profile FE, ingredient consumption and daily consumption history rounding inconsistency.

+++ frontend/src/app/consumption/TodaysConsumptionHistory.jsx
@@ -78,7 +78,7 @@ export default function TodaysConsumptionHistory({ items }) {
         >
           <div>{item.Ingredient}</div>
           <div>{item.Weight}</div>
-          <div>{Math.round(parseInt(item.Calories))}</div>
+          <div>{Math.round(parseFloat(item.Calories))}</div>
           <div>
             <button
               className="text-red-500 hover:text-red-700"
